[#120] refactor: (PR 반영) Service Interface 삭제, ErrorMsg static으로 변경 및 이름 변경 - MemberService는 Interface가 따로 필요하지 않을 것 같아 삭제
